# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Travel Plan API
  description: |
    API documentation for the Travel Plan microservices platform.
    
    ## Authentication
    All API endpoints (except `/auth/login` and `/auth/register`) require a valid JWT token.
    Include the token in the `Authorization` header: `Bearer <token>`
    
    ## Error Handling
    All errors return a consistent response format:
    ```json
    {
      "success": false,
      "message": "Error description",
      "data": null,
      "timestamp": "2024-01-01T12:00:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: Travel Plan Support
    email: support@travelplan.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server (Kong Gateway)
  - url: https://api.travelplan.com
    description: Production server
...
paths:
...
  /travels:
    get:
      tags:
        - Travels
      summary: Get all travels
      operationId: getAllTravels
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TravelStatus'
      responses:
        '200':
          description: List of travels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelListResponse'

    post:
      tags:
        - Travels
      summary: Create a new travel
      operationId: createTravel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTravelRequest'
      responses:
        '201':
          description: Travel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelResponse'
  /travels/{id}:
    get:
      tags:
        - Travels
      summary: Get travel by ID
      operationId: getTravelById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Travel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelResponse'

    put:
      tags:
        - Travels
      summary: Update travel
      operationId: updateTravel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTravelRequest'
      responses:
        '200':
          description: Travel updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelResponse'

    delete:
      tags:
        - Travels
      summary: Delete travel
      operationId: deleteTravel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Travel deleted

  # Payment endpoints
...
  /payment-methods:
    get:
      tags:
        - Payment Methods
      summary: Get all payment methods
      operationId: getAllPaymentMethods
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of payment methods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodListResponse'

    post:
      tags:
        - Payment Methods
      summary: Add a new payment method
      operationId: createPaymentMethod
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentMethodRequest'
      responses:
        '201':
          description: Payment method created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponse'
  /payment-methods/{id}:
    delete:
      tags:
        - Payment Methods
      summary: Delete payment method
      operationId: deletePaymentMethod
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Payment method deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TravelStatus:
      type: string
      enum:
        - DRAFT
        - PLANNED
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED
    Travel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
        currency:
          type: string
        status:
          $ref: '#/components/schemas/TravelStatus'
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/Destination'
        createdAt:
          type: string
          format: date-time
    Destination:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        country:
          type: string
        description:
          type: string
    CreateTravelRequest:
      type: object
      required:
        - userId
        - title
        - startDate
        - endDate
      properties:
        userId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
        currency:
          type: string
          default: EUR
    UpdateTravelRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        budget:
          type: number
        status:
          $ref: '#/components/schemas/TravelStatus'
    TravelResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Travel'
    TravelListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Travel'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'

    # Payment schemas
...
    PaymentType:
      type: string
      enum:
        - CREDIT_CARD
        - DEBIT_CARD
        - STRIPE
        - PAYPAL
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/PaymentType'
        provider:
          type: string
        lastFourDigits:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        isDefault:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreatePaymentMethodRequest:
      type: object
      required:
        - userId
        - type
        - provider
      properties:
        userId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/PaymentType'
        provider:
          type: string
        cardNumber:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        isDefault:
          type: boolean
          default: false
    PaymentMethodResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/PaymentMethod'
    PaymentMethodListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/PaymentMethod'

    # Common schemas
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...